initial checkin of Basic, AWK, and shell code
The mc* scripts are for working with PayPal Extras MasterCard credit card statements.  Using Adobe
Reader, the PDF file can be saved in text format.  From the text files, the scripts generate files
in comma-separated values format that can be imported into a spreadsheet.

The misc* scripts are for working with miscellaneous expenses that might be compiled manually as
opposed to downloaded from a web site, and are similar to the mc* scripts but are flexible to allow
for a variable number of fixed columns on the left of the transaction data.  Sometimes a transaction
is prefixed by only one date, sometimes by two dates, and as in the case of the MasterCard data by
two dates as well as a unique reference number.  The number of fixed columns can be passed as a
parameter to the AWK script.

The accounting macros are written in LibreOffice Basic for use with LibreOffice Calc.  Transaction
data spanning multiple columns and transactions spanning multiple rows can be categorized and those
categories totaled using the macros.  The programming interface is in the spreadsheet itself as much
as possible so that a user need not understand LibreOffice Basic in order to use the spreadsheets to
calculate finances.